+++ app/src/main/java/org/twinlife/twinme/ui/spaces/OnboardingSpaceActivity.java
@@ -288,7 +288,7 @@ public class OnboardingSpaceActivity extends AbstractOnboardingActivity {
         textPaint.setTextSize(Design.FONT_MEDIUM32.size);
         textPaint.setTypeface(Design.FONT_MEDIUM32.typeface);
 
-        int textWidth = (int) (Design.DISPLAY_WIDTH - (Design.ONBOARDING_TEXT_MARGIN * 2));
+        int textWidth = Design.DISPLAY_WIDTH - (Design.ONBOARDING_TEXT_MARGIN * 2);
         Layout.Alignment alignment = Layout.Alignment.ALIGN_NORMAL;
         StaticLayout staticLayout = new StaticLayout(message, textPaint, textWidth, alignment, 1, 0, false);
         return staticLayout.getHeight();
